rename: ui 폴더 component 하위로 변경
